moving and testing Password Service
