+++ src/HomeNet.Infrastructure/HomeNet.Infrastructure.csproj
@@ -13,6 +13,8 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" />
     
+    <PackageReference Include="BCrypt.Net-Next" />
+
     <PackageReference Include="Dapper" />
     <PackageReference Include="Npgsql" />
     <PackageReference Include="SqlKata" />
